requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-environ==0.4.5
django-allauth==0.42.0
django-filter==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
